extract onKeyPress to own module for mocking
